Add .west to zephyr-src output
By doing so we allow the whole derivation to be recognized as a West
workspace.

The current problem is that West always opens the config file with write
permissions, which it cannot do as the Nix store is read only. I do not
think it actually needs those permissions, but we will see.

Signed-off-by: Niklas Wimmer <mail@nwimmer.me>