Rename the West manifest repository source
If the source's name is 'zephyr' some internal West logic activates that
cause it to try writing something into the workspace's config file.
Because that config file is located in the Nix store however, hardly any
West command can be used as a result (update works, which is why this
went unnoticed at first).

The West code that causes this behavior can be found here:
https://github.com/zephyrproject-rtos/west/blob/v0.14.0/src/west/app/main.py#L415

The removal of this behavior is tracked as part of
zephyrproject-rtos/west#246

Signed-off-by: Niklas Wimmer <mail@nwimmer.me>