Install python dependencies at shell creation time
By doing so the reproducability is reduced, because not all Python
dependencies are pinned to one specific version. In practice though, the
whole Zephyr ecosystem does that anyways, so even if something breaks,
it would expect it to be noticed and fixed quite fast.

Another downside is that the dependencies are no longer part of the
shell derivation which means they cannot be cached. This will probably
increase CI times by a few short minutes.

The advantage is that the Python dependencies do not need to be
specified by hand. Previously, only the base packages were available,
exactly because specifying everything by hand is way to tedious.
Also, if the Zephyr version is updated, the Python dependencies
are updated as well without any additional work.

Overall the change is net positive. Should CI times or the
reproducability become a problem something else needs to be thought of.

Signed-off-by: Niklas Wimmer <mail@nwimmer.me>