Fix toolchain setup
As it turns out I was using another toolchain for testing without
noticing which is why I originally thought the previous setup works.
Turns out it did not.

Zephyr now also kind of expects that it can find the host-tools at paths
which do not exist, but as long as noone tries to actually invoke them,
it does not matter.

Signed-off-by: Niklas Wimmer <mail@nwimmer.me>